ftp模板

+++ newplan_core/src/main/java/club/zby/newplan/controller/ftpserviceclient/ControllerFtp.java
@@ -48,9 +48,9 @@ public class ControllerFtp {
 
     @ResponseBody
     @ApiOperation(value="数据库文件列表", notes="数据库文件列表测试")
-    @PostMapping(value = "basefileList")
-    public Result basefileList() {
-        return ftpClient.basefileList();
+    @GetMapping(value = "basefileList/{page}")
+    public Result basefileList(@PathVariable("page") Integer page) {
+        return ftpClient.basefileList(page);
     }
 
 
@@ -58,11 +58,11 @@ public class ControllerFtp {
     @ApiOperation(value="下载文件", notes="文件下载测试")
     @PostMapping(value = "downFile")
     public Result downFile(@RequestParam("fileName") String fileName, @RequestParam("localPath") String localPath){
-//        String status = (String) request.getAttribute("status");
-//        System.out.println(status);
-//        if("404".equals(status) || status == null){
-//            return new Result(false, StatusCode.LOGINERROR,"登录异常",null);
-//        }
+        String status = (String) request.getAttribute("status");
+        System.out.println(status);
+        if("404".equals(status) || status == null){
+            return new Result(false, StatusCode.LOGINERROR,"登录异常",null);
+        }
         return ftpClient.downFile(fileName,localPath);
     }
 
@@ -71,11 +71,11 @@ public class ControllerFtp {
     @ApiOperation(value="删除文件", notes="文件删除测试")
     @DeleteMapping(value = "delFile")
     public Result deleteFile(@RequestParam("fileName") String fileName){
-//        String status = (String) request.getAttribute("status");
-//        System.out.println(status);
-//        if("404".equals(status) || status == null){
-//            return new Result(false, StatusCode.LOGINERROR,"登录异常",null);
-//        }
+        String status = (String) request.getAttribute("status");
+        System.out.println(status);
+        if("404".equals(status) || status == null){
+            return new Result(false, StatusCode.LOGINERROR,"登录异常",null);
+        }
         return ftpClient.deleteFile(fileName);
     }
